Move HttpRequester and tests
