+++ lsp/core/aws-lsp-core/package.json
@@ -18,6 +18,8 @@
         "@types/mocha": "^10.0.1",
         "chai": "^4.3.7",
         "chai-as-promised": "^7.1.1",
-        "mocha": "^10.2.0"
+        "mocha": "^10.2.0",
+        "sinon": "^15.1.0",
+        "ts-sinon": "^2.0.2"
     }
 }
